+++ src/main/java/dev/vijay/productservice/services/ProductService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface ProductService {
 
-    Products getSingleProduct(int id);
+    Products getSingleProduct(long id);
     List<Products> getAllProducts();
     Products createProduct(int id , int price , String name, String description,Products products );
 }
